example/pyproject.toml
[project]
name = "yeti-demo"
version = "0.1.0"
description = "Demo clients for Yeti pipeline"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.0.2",
    "requests>=2.31.0",
    "hypothesis>=6.92.0",
    "pymongo>=4.6.0",
    "redis>=5.0.0",
]
